services/auth-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
werkzeug==3.1.3
httpx==0.27.2
pybreaker==1.2.0
